legacy/requirements_old.txt
PyQt6>=6.4.0
pyqtgraph>=0.13.0
pypylon>=2.0.0
opencv-python>=4.8.0
pyserial>=3.5
numpy>=1.24.0
pandas>=2.0.0